// src/app/models/ProyeccionVentaDet.model.ts
export class ProyeccionVentaDet {
	idItem: number;
	alNave: string;
	viaje: string;
	eta: Date;

	ca2SdNoFe = 0;
	ca2SdFe = 0;
	ca4SdNoFe = 0;
	ca4SdFe = 0;
	ca4ShNoFe = 0;
	ca4ShFe = 0;
	ca4RhFe = 0;
	ca4RhNoFe = 0;

	ca2SdNoFePick = 0;
	ca2SdFePick = 0;
	ca4SdNoFePick = 0;
	ca4SdFePick = 0;
	ca4ShNoFePick = 0;
	ca4ShFePick = 0;
	ca4RhNoFePick = 0;
	ca4RhFePick = 0;

	fgFile: string;
}